Move `chat` store into module
