+++ app/src/renderer/store/chat/actions.js
@@ -1,4 +1,4 @@
-import * as api from '../api'
+import * as api from '../../api'
 import * as types from './mutation-types'
 
 export const getAllMessages = ({ commit }) => {
