+++ app/src/renderer/store/chat/getters.js
