+++ app/src/renderer/store/index.js
@@ -1,44 +1,13 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
-import * as getters from './getters'
-import * as actions from './actions'
-import mutations from './mutations'
-// import createLogger from '../../../src/plugins/logger'
+import chat from './chat'
 
 Vue.use(Vuex)
 
-const state = {
-  currentThreadID: null,
-  threads: {
-    /*
-    id: {
-      id,
-      name,
-      messages: [...ids],
-      lastMessage
-    }
-    */
-  },
-  messages: {
-    /*
-    id: {
-      id,
-      threadId,
-      threadName,
-      authorName,
-      text,
-      timestamp,
-      isRead
-    }
-    */
-  }
-}
-
 export default new Vuex.Store({
-  state,
-  getters,
-  actions,
-  mutations,
+  modules: {
+    chat: chat
+  },
   plugins: process.env.NODE_ENV !== 'production'
     ? []
     : []
